Visualizerを分けた
